Bug 1132952 - Process and some other platform services are not listed under group creation dialog search filter
An issue with the runs-inside feature in plugin descriptors.

Now showing resource types under all their parents.

For example, the Process service will appear under Linux Platform, Windows Platform, ... etc.

On the flip side: this change might lead users to think that they are, for instance, selecting all Process services which run under a Windows Platform.
But this is not true. They are actually selecting *all* Process services, regardless of the parent platform.