[BZ 1135034][BZ 1135107] Disallow patching on unsupported EAP versions.
This is done on 2 fronts. First, there's a new "supportsPatching"
attribute in the plugin config of standalone and host controller to tell
whether the resource explicitly supports patching or not. This is to
support servers that don't support patching even though the patch
command is available.

As a second "wave", the bundle handler checks for availability of the
patch command before trying to perform a patch operation and returns
a meaningful error message if it fails to find it.
This is to support graceful failure on servers that somehow fail to
report whether they support patching correctly.