[BZ 1135630] add metrics_idx table and remove cache tables
This is a big commit that involves a lot of clean up due to the removal of the
metrics_cache and metrics_cache_index tables. This commit also includes a lot
of work updating the aggregation code to use the new metrics_idx table. All of
the tests in AggregationTests.java pass with this commit as well.

Conflicts:
	modules/enterprise/server/server-metrics/src/main/java/org/rhq/server/metrics/MetricsDAO.java
	modules/enterprise/server/server-metrics/src/test/java/org/rhq/server/metrics/MetricsDAOTest.java
	modules/enterprise/server/server-metrics/src/test/java/org/rhq/server/metrics/invalid/InvalidMetricsManagerTest.java