[BZ 1069547] Support shutdown of domain controller in ServerControl
This fixes the patching domain controllers which failed if restart
was demanded during the deployment.

Also optimized the reading of host.xml (to find the host of the domain
controller in the domain) to use SAX parser instead of DOM. This is
much lighter on resources and should also be much quicker because we
only need to read first couple of lines of the file.