[BZ 1136945] Purge/revert of EAP patch bundles works more sanely.
To implement this reasonably, we started tracking the state of
patch bundles deployed through RHQ in external file, located under
the target Wildfly/EAP server's .installation/.rhq subdirectory (the
.installation directory is used by Wfly/EAP itself to keep track of
patching, so it is a logical place to add patching related information).

During deployment we determine the effective set of patches that were
applied from the patch file (by comparing histories before and after
deployment). We then use this information to revert/purge only the
patches that were deployed using RHQ. This information is stored per
destination in the above mentioned metadata directory (so that multiple
destinations per EAP are supported (even though such arrangement is
non-sensical)).

In another words all patches applied before a first patch bundle was
deployed through a destination are retained after a bundle purge and
the revert or purge of the bundle will fail or warn if the history
of patch deployment is unexpected (due to external patch rollbacks
or applications).

The revert/purge will fail if no patches can be rolled back due
to the state of the history having a different "tip" than expected.
A warning is emitted to the audit log if some of the patches that
were expected to rollback couldn't be because of some other change
in the patch application history.

There situations will only exist if the EAP server is patched through
other means than RHQ after it has been patched using RHQ thus creating
and inconsistency between what RHQ sees as the deployment history of
the destination and the real state on the EAP resources.