[1071379] CTRL+C doesn't shutdown server when running in console mode on Windows Rhqctl was actually the program running in the foreground when the 'rhqctl console --server' command was issued. The JSW and Server processes were also running, but in the background.  When issuing a CTRL-C it was actually rhqctl that was getting interrupted and killed. The CTRL-C was not reaching the wrapper, or the server.  This got a little infuriating because even if you manually killed the Server, the wrapper would detect the "crasg" and then spin up a new server process :)  So, to recover you needed to manually kill the wrapper.exe process, and then manually kill the RHQ Server process.
With the fix we now launch a new cmd window for the Server (asynchronously,
via the Windows start command) which allows rhqctl to complete and exit.
The wrapper is now running in the foreground of the new window, and CTRL-C
is handled as expected by JSW (shutdown), as is double CTRL-C (immediate
shutdown).

Analogous work done for --storage.  Note that we don't support 'rhqctl
console --agent' and already direct users to run rhq-agent.bat directly.