[BZ 1069547] Adding "Active Patches" trait to Host ctl & Standalone svr
As such, the user now has the overview of the patches applied to the
individual Wfly/EAP servers in their env. This complements the bundle
subsystem from which the same information could be deduced, too,
under the assumption that RHQ server would be the sole "issuer" of
the patches to the servers.