[BZ 1114200] fix handling of late data
My previous commits that involved setting the column timestamp on inserts had
to be reverted. When aggregating late data, it is possible to end up with a
clock tie (See https://wiki.apache.org/cassandra/FAQ#clocktie for details). In
those situations the columns might not get overwritten.

The key thing is to ensure that we do not try to recompute aggregate metrics
when some or all of the source data has already expired, leading to new,
incorrect data. I have put changes in place that should prevent this from
happening.

This commit also includes a change to IndexIterator. It will now continue
loading pages if there is a read timeout. Previously, the exception would go
uncaught causing data aggregation to abort early.