Bug 1125423 - Fuse Fabric Groups Plugin cannot be enabled
Generally when there is an attempt to enable server plugins we now first
check it's configuration by looking for required properties (in ConfigDef)
and making sure there's valid value in configuration. If something is
missing, it means plugin is not configured and cannot be enabled.
ServerPluginManagerBean thows this using
PluginConfigurationRequiredException that is then presented to user.