Bug 1146583 - "Ignore" button is disabled always under "Discovery Queue"
Previous logic of enabling/disabling "Ignore" button tried to prevent
situation, that user could ignore a NEW platform and/or it's child. Previous
logic would not work in case user also selected at least another resource
that "could" be ignored (a child of imported platform).

Now user can ignore MEW platform as well or it's child. If platform is
IGNORED it's not possible (at least from UI) to import it's NEW children.