[1139735] Add undo of the storage node version stamping in case of an upgrade failure invoking undo logic.   The current thinking is that this is not necessary for the server stamp since it happens at the end of server upgrade.
Also, deprecate the upgrade --start option and make it do nothing if specified.
This is because upgrade is a multi-step process and as such an automatic
start does not make much sense, and more likely will cause problems. Start
should only be done manually, after all upgrade steps are complete.