Bug 1154678 - [EAP6] - Newly created datasource is not enabled even when you have set Enabled=true -- user should be at least warned
When the datasource was created, the 'enabled' attribute was supplied to the EAP6 management interface.
But setting the attribute to true does not enable the datasource. An extra server reload is needed.
To enable a datasource without a server reload, the 'enable' operation needs to be called.

So now the DatasourcesComponent class:
-> no longer supplies the 'enabled' attribute to the CreateResourceDelegate
-> reads the state of the datasource after creation
-> invokes the 'enable' or 'disable' operation (as needed) after creation

(cherry picked from commit d5e65f48de47b1c3258277e5e3e7d16881b679fc)
Signed-off-by: Libor Zoubek <lzoubek@redhat.com>