[BZ 1160775] - RFE Adding a way to show/hide RHQ subsystems in UI
Current UI is overcomplicated. RHQ does lots of things and
different users use it for different purposes. Although many
subsystems are interconnected and may depend on each other
(e.g. metrics and alerts). This is not the case for all of them
and in the sake of simplicity some of them could be hidden in
UI if user wants to.

This PR allows to hide following subsystems/features in UI:
 * content
 * reports
 * administration
 * events
 * operations
 * alerts
 * config
 * drift

NOTE: This change is only about hiding it in UI, the subsystems
are still working and can be acessed via REST, CLI or direct
URL.

The subsystems can be shown/hidden on the user page
${host}/coregui/#Administration/Security/Users/${userId}
and this information is saved to the user preferences.

It is orthogonal to the access rights that are specified for
roles. Each user may decide what to hide to make the UI
cleaner for him/her.

it's 1 squashed commits of following:
 This is the 2nd commit message:

[BZ 1160775] - RFE Adding a way to show/hide RHQ subsystems in UI - fixing FF issues and changing the switches according to UX proposal

 This is the 3rd commit message:

[BZ 1160775] - RFE Adding a way to show/hide RHQ subsystems in UI - i18n and improving CSS for the switches

 This is the 4th commit message:

[BZ 1160775] - RFE Adding a way to show/hide RHQ subsystems in UI - rewriting it to use the bootstrap-switch library; messages

 This is the 5th commit message:

[BZ 1160775] - RFE Adding a way to show/hide RHQ subsystems in UI - final tunning of the timing