Bug 1168890 - Remove InventoryStatus.DELETED resource state
DELETED InventoryStatus makred as deprecated and it is not "mentioned" in
server/agent code anymore, this has a consequence in DeleteResourceHistory,
which now refers to parent resource and no longer refers the deleted
resource - instead it refers to resource type, resource name and resource
key (for history purpose, so we know what was deleted). This has another
consequence in DeleteResourceResponse from client-api which now has to hold
the deleted resource id.