BZ-1203469: Fail to manually import JBoss EAP based products when product slot is unknown
If a JBoss EAP 6/AS 7 instance does not include a bin/product.conf file,
it is assumed to be an AS 7 instance. However, if it does include a
product.conf file but the slot value is unknown by the as-7 plug-in, an
attempt is made to determine its type from its home directory name. If
this does not work, discovery is aborted and the JBoss EAP 6 / AS 7
instance can not be imported into inventory.

Even when an AS7 resource is manually added, this same problem occurs.

This fix allows a manual import to be successful even if the product
type can not be determined. A warning is now logged to identify that an
unsupported product was found and the product type falls back to AS.