BZ-1203469: Fail to manually import JBoss EAP based products when product slot is unknown
If a JBoss EAP 6/AS 7 instance does not include a bin/product.conf file,
it is assumed to be an AS 7 instance. However, if it does include a
product.conf file but the slot value is unknown by the as-7 plug-in, an
attempt is made to determine its type from its home directory name. If
this does not work, discovery is aborted and the JBoss EAP 6 / AS 7
instance can not be imported into inventory.

Even when an AS7 resource is manually added, this same problem occurs.

This fix allows a manual import to be successful even if the product
type can not be determined. A warning is now logged to identify that an
unsupported product was found and the product type falls back to AS.

(cherry picked from commit 0226e3c1164a696da127ba7a1eaf4d80c013216c)

Conflicts:
	modules/plugins/jboss-as-7/src/main/java/org/rhq/modules/plugins/jbossas7/BaseProcessDiscovery.java

  See [BZ 1300727] for more details, but this commit squashes the following four commits and cherry-picks from 3.1.x->3.3.x
    0b4745e8a6c48943fcb821c5cd557bc7f03678bd
    5bcfe335ca257bcc142d9ac74103e1b5f911b61e
    0a9a4eee63ca268da106bd6d20b1c0dc4252e645
    181b92a192b9b09c96d2ebaf62424944a969880c

Conflicts:
	modules/plugins/jboss-as-7/src/main/java/org/rhq/modules/plugins/jbossas7/BaseProcessDiscovery.java