BZ-1203469: Fail to auto discover EAP based products when product slot is unknown
If a JBoss EAP 6/AS 7 instance does not include a bin/product.conf file,
it is assumed to be an AS 7 instance. However, if it does include a
product.conf file but the slot value is unknown by the as-7 plug-in, an
attempt is made to determine its type from its home directory name. If
this does not work, discovery is aborted and the JBoss EAP 6 / AS 7
instance can not be imported into inventory.

This fix allows the product type to fallback to a generic AS instance in
the event its product slot is unknown and a the home directory name does
not include a product key in its name.

(cherry picked from commit 29b503b5cfd3f9a99c1a8c10bac23b1c272bd0af)