Bug 1173164 - Globally uncaught Exception shown in the JBoss ON UI when trying to display orphaned operation schedule
Fixed by removing operation schedules (and unscheduling it) when the user
does not exist anymore. This removal does not happen at the time when user
is deleted, it happens when either - user want's to see such schedule (which
refers to missing user) or when scheduled job is about to run (in this case
it is canceled).