[BZ 1252142]  Creation and group association of bundle fails with timeout followed by IllegalArgumentException The original fix commit for this problem is valid (1990e7d65).  It increased the RPC timeout from the default, 30s, to 10m in order to provide ample time for large distribution files to be processed server-side.
This uncovered an unrelated issue with the way our RPC timeouts are
handled. In short, gwt keeps only one instance of each rpc service even
though you can reassign [custom] request builders, which we do, to apply
our timeouts to the service. But each time code looks up the service it
reassigns the timeout to it's needs.  So if you lookup the service and
assign a 10m timeout it can be 'overwritten' before it's actually used. And
that was what was happening. I believe it could behave differently
on different systems but in fact, on my system, it was overwritten twice
between two consecutive lines of code, the first looking up the bundle
service and the econd making the call.

The reason is that we don't control the order of everything going on in gwt.
For example, several classes initialized the bundle service as a
class-field.  The initialization could happen at unexpected times.  Instead,
by chaining every bundle-service-lookup to the service call, the problem has
gone away.  I particularly suspect the class-level decls in the DataSources,
but we had plenty of advance initializations going on.

This problem could may still affect other code using other services. This
commit is for the bundle service only, other changes were out of scope.