BZ 1264395 Agent is not started correctly on RHEL 7.1 OpenJDK 1.6.0_34 The issue happens during schema validation (of the platform plugin) and not during general parsing.  If the validation fails optionally re-parse the plugin descriptor skipping the xsl validation if we're running a problematic version of java.
The problematic java versions are controlled with the org.rhq.xsl.validation.skip
sysprop.  If unset it defaults to all OpenJDK 1.6 versions.  It is
unbounded because the problem is in the most recent released versions. If
a version is released that fixes the issue, we could update the pattern
to be bounded to the exact problematic versions.