Bug 1246083 - CLI operations don't work when Native Local Authentication is enabled and Native management API host is set to 0.0.0.0
Updated discovery class so that when '0.0.0.0' is discovered, '127.0.0.1' is used
Updated component class to log a warning if native host is set to '0.0.0.0'
Changed defaults from 'localhost' to '127.0.0.1' in HostConfiguration class to avoid troubles when localhost is configured to something else than loopback address

Note that, with these changes, new servers bound to '0.0.0.0' will no longer have this address in the resource name, but '127.0.0.1' instead.
Resource keys are not impacted. Already inventoried EAP6 resources are left untouched.